Re-create neural dynamics and tested
