+++ CMakeLists.txt
@@ -40,7 +40,7 @@ option(GUROBI_ROOT "Path to Gurobi installation" "")
 set(GUROBI_ROOT $ENV{HOME}/.local/lib/gurobi1103/linux64)
 
 # LibTorch Configuration
-option(CDDP_CPP_TORCH "Whether to use LibTorch" ON)
+set(CDDP_CPP_TORCH "Whether to use LibTorch" ON)
 option(CDDP_CPP_TORCH_GPU "Whether to use GPU support in LibTorch" ON)
 set(LIBTORCH_DIR $ENV{HOME}/.local/lib/libtorch CACHE PATH "Path to local LibTorch installation") # FIXME: Change this to your local LibTorch installation directory
 
@@ -198,7 +198,7 @@ set(cddp_core_srcs
   src/cddp_core/cddp_core.cpp
   src/cddp_core/asddp_core.cpp
   src/cddp_core/logddp_core.cpp
-  src/cddp_core/torch_dynamical_system.cpp
+  src/cddp_core/neural_dynamical_system.cpp
 )
 
 set(dynamics_model_srcs
